feat(alarm): viewportStore -> useHeaderStore
 - viewportStore를 useHeaderStore로 변경
 - 알람 State추가
 - install immer

Resolves #53
