+++ package.json
@@ -21,6 +21,7 @@
     "@tiptap/pm": "^2.11.5",
     "@tiptap/react": "^2.11.5",
     "@tiptap/starter-kit": "^2.11.5",
+    "immer": "^10.1.1",
     "next": "15.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
